Add options to onion skinning so user may define the number of frames shown.
Add options in Edit > Preferences > Tools so that the user may define how many previous and next frames are shown when previous and/or next onion skinning is toggled on. The user may also define a maximum opacity value and a minimum opacity value, which will be gradated between all of the frames shown. Replaces the previous options here which allowed the user to define the opacity values for the next or previous 3 onion frames explicitly; these settings were broken and did not work.