Do not show a second warning dialog if file open fails
From my understanding, newDocument needs to be called if file open
encounters an error or else the program could be left in a bad state.

Users are already prompted about unsaved changes when they try to
open a file, so there should be no additional risk of loosing
unsaved work as a result of these changes.