Fix rename optimization during save
During saving, there was some code to check if a frame changed
position but was not modified so that a simple file rename could
be used instead of a full image render and write.

However, there were many bugs that prevented this from working
properly. I hope I have fixed them all. The most major one was
caused by changes to the image loading, which would cause
all frames to be loaded into memory during save. This fix will
reduce memory consumption and increase the save speed by fixing
this.