Fix audio volume in movie export
Amix has an automatic dropout that adjusts the volume to 1/n,
where n is the number of sounds currently playing. This isn't
necessarily a bad thing to do, but it doesn't work for us because
adelay pads the beginning with silence rather than offsetting the
clip. Also I don't think qt does this so we should try and make
it sound consistent with the preview playback. The fix is to
use amerge + pan instead.