Fix various bugs related to previous drawing action
notes: Nothing that hasn't been saved in the backup should be checked in an undo or redo action as the content may have been modified later.

other things:
I had a revelation about bitmap transformations \o/, so much of the transformation stuff that was added previously is now gone again.