Improve color wheel square precision
There were a few things off when rendering the square section of
the color wheel, and the picker inside of it. The picker should
now be exactly centered by the cursor, and should not extend past
the boundaries of the square. The color selected may also be more
accurate, although it still does not perfectly match the RGB data
from the Digital Color Meter utility (macOS). The issue is
probably in the hsv to rgb conversion, rounding, or gamma.