Fix selection expansion bug when moving
When moving a selection, it would sometimes grow by a pixel in
width and/or height. While it makes sense to want to create a
selection that fully encompasses the selected area as
toAlignedRect does, it results in some confusing and unintuitive
results when adjusting. Rounding as toRect does seems much more
natural when you are using it, and doesn't change the size of your
selections when dragging.