Refactor canvaspainter out of scribblearea
CanvasPainter now resides in canvasmanager, this should make it easier to write tests that would otherwise require scribblearea.

Additionally I've also added a keyframemanager, currently it doesn't do much but I felt it made sense for stuff like getting the current keyframe on a layer, which was previously defined in Scribblearea.