Fix restoring camera key would crash with malloc
I don't quite understand why it crashes here but the issue is basically that loadKey is called, setting the pointer of a keyframe which moments later is then freed while still being in use and thus crashes... Luckily it's not needed at all here anyway, since we can simply assign the camera to the current camera.

Additionally I've added a proxy to update the current view, otherwise the motion is not shown immediately.