TransformElement: backup selection flip
The implementation of this made me realize an underlying problem with the previous transform calculation... since the scale is never handled, but simply calculated from the rect.. which will always be the same. The result of that meant that flipping the selection would reset after moving. Additionally there's also no reason to apply the selection transformation.

- Remove merge from transformElement.. this means backup pieces of this element are never merged now.
+ Add ScaleX and ScaleY to selectionmanager.
+ Improve various transform descriptions