Make ffprobe optional by default and don't include by default
I'm not in favor of relying upon ffmpeg for retrieving the
duration. It is more difficult to parse and highly suseptible to
changes in format between ffmpeg versions and possibly
locatalizations. I have made the ffmpeg duration code a fallback
if ffprobe is absent, that way if it does cause issues, we can
just tell users to throw the ffprobe binary in the plugins folder
and it will just work.