Modify handling of mouse and tablet events
These changes fix a few bugs with mouse and tablet inputs,
including:
- Program was left in inconsisent state (mMouseInUse=true but
mouse is not actually in use) when trying to draw on a hidden
layer with a mouse.
- After fixing the previous bug, when trying to draw on a hidden
layer with a tablet, the message would come up, but then it would
draw on the hidden layer anyway.
- Many actions which are rightfully blocked when drawing with a
mouse were allowed when drawing with the tablet (ex. zooming with
mouse wheel or using shortcuts)