Fix partial file wipe issue #1280
The problem is that the temporary location for bitmaps was
a relative path. This means that where exactly it would move
the images to while saving is os dependent. On mac, if this
was run from terminal it would be the current working directory
which usually does not result in a problem. However when running
the app graphically, it is relative to the root directory, which
is not writeable. I did not debug this on Windows, but I suspect
that it was also relative to the root directory, or at least
some other unwritable directory. It now keeps the images within the
data dir which should always be writeable.