Fix polyline tool regression
The polyline tool was not correctly displaying while drawing. This
was probably broken by the changes to the frane caching in #1267.

I also did a little bit of refactoring to implement this more
cleanly. Tools now report if they are modifying the canvas with
isActive, rather than ScribbleArea trying to figure that out.
I also changed the right click action to set the temporary tool to
the hand tool rather than forwarding move event to the hand tool.
This is the same way that the middle mouse button works and is
slightly better because it will show the hand tool cursor.