Fix path issues in API docs generated on Travis CI
The current documentation on our website has some issues related to paths. For
example, some pages contain references to the directory structure in the build
environment (even in the URLs), and some links in the Markdown pages from docs/
incorrectly lead to the Markdown sources of other pages, instead of the rendered
versions.

After some investigation, it turned out that the cause of this issue is that
the Doxyfile we use for Travis specifies absolute paths based on the
TRAVIS_BUILD_DIR variable. To get correct results, it is necessary to use
relative paths (and therefore run Doxygen from the repository root).

As a side effect of this change, there are now so few Travis-specific Doxygen
settings left that it becomes feasible to generate the Doxyfile on the fly. This
saves us the trouble of having to keep two slightly different configuration
files in sync with each other.