Fix Recent File list wipe issue
The recent file list was being wiped on startup because we are now
opening a project on startup (templates) which saves an empty
recent file list before the old one can be loaded. Additionally,
it was saving every time it clears the list, which is not a good
idea since that can be used internally before making other changes.