Round selection to the nearest pixel when on the bitmap layer
This was fixed and then reintroduced by #1223. All points of a
selection need to be a whole number right now. Otherwise you have
borders drawn ambiguously between pixels, and obscure issues
when the values are rounded differently. For instance, flipping a
selection and then moving it could in some circumstances leave
behind a row/column of pixels.