Pencil2D 0.6.5 Release Changelog [WIP]
Since Pencil2D last stable release was deployed over a year, due to the huge amount of commits the categorization style has not yet been implemented and instead it's being listed in chronological order from oldest to latest commit. There are some commits that are difficult to describe as they contain several fixes of heterogeneous nature and have retained their merge-to-master automated title descriptions, but this will require some discussion as to how we want them represented.