Only push frames that are connected to new frame.
We don't want to push all frames ahead since it could create unwanted side effects like making the animation come out of sync because all frames ahead would be pushed + 1. This is still partly true now but the effect of it will only be limited to the frames that are connected to the new frame, minimizing the damage.