Don't add frame unless it's the current pos.
I've always felt it a bit cumbersome that when you want to move to a frame, you have to click on the frame indicator row instead of just the frame you want on the layer. This proposal improves on that by allowing you to double-click on any given layer layer to jump right to it.