Re-add empty input path check to CLI exporter
The input argument can be left out for GUI launching, so this
cannot be enforced by commandlineparser. While Editor::openObject
does catch instances of empty filepaths, it produces an error
that the file cannot be found, which can be confusing if the
actual issue is that no file was specified in the first place.