#1584 - Minor refactoring
At some point I would like to have all camera related logic in its own CameraTimelineCell class so that we don't pollute the general timeline cell class with Camera Layer specific code. This requires some work though as the timelinecells class is not built for that kind of setup currently.