Fix backup and undo relying on layer index to undo..
This won't do as we don't backup layer movement or anything. If you create three drawings on layer 1, then go to layer two, swap the layers and undo.. the logic will (based on the index) try to undo on the wrong layer...

I've made just enough changes to make the undo/redo feature work for filling on a different layer. There might be cases i've missed but I don't want to put too much energy into this because of the undo/redo rewrite...