Fix scrubber would still be shown as if mouse cursor was under it.
When leaving the widget..
One way to produce this was to put the mouse cursor on the scrubber, then move it along the y axis, until the cursor is outside the widget. The result would be that the scrubber would still be shown as if the cursor was there, until it enters the widget again.