Fix follow camera import image option
There were two issues with this feature. First was that it was
importing with the camera view transform, but the camera rect is
actually transformed by the inverse of that transformation.

The second issue was an occasional off-by-one error in the position
caused by using the more accurate qtransform over the rounded qrect
that the camera rect is based off of.