Use custom read function to support std::istream in miniz
This makes it possible to do cross-platform writing of data to the
zip archive without writing the data to the hard drive first.

Currently this is used to write the special mimetype file, but
without having looked into it, it could be useful for faster writing
of the main xml file or even frame data already in memory (cached
or dirty).