Fix defaulting to non-primary language in some cases
If the primary language of the system is English, but there is a
secondary language (ie. multiple languages in
QLocale::system().uiLanguages()), then the English locale will not
match any translation files. It will then check the secondary
language, and if that does match a translation file, then that
language will be used instead of English. This issue is fixed
simply by having an English translation file.