Implement path control point move state
This commit introduces a new property for the camera so that we can check whether a point was moved or not, which enables us to:
1. when the handle hasn't been moved and you move it, it will keep resetting (eg. center between the previous and current frame), until the mouse released event has fired.
2. After the path has been moved, the curvature will be kept, which means that you will be able to move the camera without resetting the path.

Additionally!! I have modified the xml so that all new path related data follows the same naming convention. This means that old projects with the "midx" and "midy" attributes will not work as expected.