Mitigation: Fix undo/redo modification not being applied
This happens because our logic assumes the current layer changes, meaning you can't apply vector transformations while you're on bitmap layer and vise versa. This will fix that, by making sure you are on the correct layer.

In the future we should figure out how we want this to behave...