Remove ability to delete first layer (camera layer)
This is a fundamental change but one that I believe we'll benefit from. Since we already have a sort of hierarchy by order just not fully established... This will set that in stone.

From this commit forth, there should always be a camera at the bottom, because we need to tie overlays and other things to a camera, likewise the bucket tool is dependent on the camera boundary. There are currently several problems with moving a camera above a bitmap or vector layer, making the software more unstable than it needs to be.