Fix view freezing when both rotating camera and view
The problem was caused by QRegion ending up cutting the transformed rect/polygon into thousand of smaller slices (anywhere between 4 with no transformation to 9000 cuts after transforming to certain angles!), make the application freeze for several seconds.

Fortunately the problem can be fixed rather trivially. We don't need to clip at all, instead we simply create a polygon that matches the look we want, in this case the passpartout effect.