Use different naming scheme for cloned frames
The old approach added a 12 character suffix to the source frame's
filename. This works fine when cloning an normal frame, but cloning
cloned frames would result in multiple suffixes, rapidly increaing
the length of the file name. While untested, I suspect this could
cause serious issues when there is a small limit on the file name
or path length (such as on Windows).