Don't force file name or modification state of cloned keyframes
Now that the backing file is properly copied, we do not need to
set duplicated keyframes as modified. If the source frame is
modified, the cloned frame will be modified as well and remain in
memory. If the source frame is not modified, then it can be copied
without loading it into memory, or unloaded after copying.