Add more details to FileManager error report (#1895)
* Fix case where saving with no undo changes would fail

* QMiniz: Simplify logic

* QMiniz: Make sure to abort when failing to compress file

The problem with not aborting here is that while one file can fail, it will be noted down but if the next file succeeds, then it will seem like the project got compressed successfully.

* Improve the error detail log for readability

And get rid of useless or confusing information.

* ErrorDialog: Add copy to clipboard

* Improve error details for readability - take 2

* Improve error details for readability - take 3

* Write a note when a backup has been made

* FileManager: Fix logic does not account for existing backup files

The logic didn't account for existing files with "backup" suffix, eg.
A folder may contain:

MyProject.pclx
MyProject.backup1.pclx

If the user opened MyProject.pclx, and failed to save it later, the logic would not count existing backup files like: "MyProject.backup1.pclx" and as such the result would be no backup being made.

* Improve layer saving diagnostics

* Fix where where project couldn't be saved because miniz tried to zip non existent files

This is caused by the new logic actually catching when a file has gone missing while saving. While it may not fix any bugs inheritly, it should make it easier to diagnose problems in the future.

* Improve diagnostics around qminiz and archiving

* Add text explaining the importance of the bug report

* Add missing space

* QMiniz: make sure to always close reader

* Layer: Add error diagnostics when failing to save one or more layers

* FileManager:writeMainXML: Slightly safer way to close file

* QMiniz: Make errors more explicit and fail quickly.

* Add error keyword to writePalette

* FileManager: close file after recovery.

* Remove indentation

* FileManager: Check for Unzipping errors

* QMiniz: Fix typo...

I'm blown away by the fact that this worked at all... wtf. The only one who reported a problem here was windows but technically the file should never have been closed on any of them.

Frightening..

* Add missing QFile::close where a file has been opened.

And move some closer to the open call to make sure it happens when the scope ends.

* ErrorDialog: shorten description

* QMiniz: Only close if we succeded to open it

.

* Remove explicit file.close because QFile already handles this

* Fix relative include path
# Conflicts:
#	core_lib/src/managers/undoredomanager.cpp
#	core_lib/src/structure/filemanager.cpp