iss-1897 - Fix view rotation effects affects various view handles
The code getViewScaleInverse assumes the m11 value is a scaling matrix but that's not guaranteed! We see the problem occur when the view has been rotated, which changes the matrix type so that it's no longer a pure scale factor. The fix is to create a new scaling matrix from the mScaling property and inverse that instead.

Simplify the logic