Fix #1873 - Fix frame not being drawn on the timeline (#1887)
* AC-1873 - Fix frame not being drawn on the timeline

The reason for this is that while we keep a list of keyframes to move and paint keyframes on the timeline, a keyframe also has its own "isSelected" state. This state is backed up through undo/redo operations which the layer::getSelectedFrames is not.

The solution is to remove isSelected from Keyframe, as there should only be one truth.

* Only deselect layer if the index has changed

* TimeLineCells: Use ForEach and return

* Cleanup unused function