Fix inaccessible top menu bar items
The menu was being blocked by empty QWidgets from the title bars.
By default, the title bar is shown and the empty title bar is hidden.
They are then shown and hidden by QDockWidget::setTitleBarWidget.

The parent widget of the titles bars was also changed to the
QDockWidget as that is more logical and assures that the title
bars are deleted if the BaseDockWidget is deleted.